backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg2-binary==2.9.10
pydantic==2.10.4
python-dotenv==1.0.1
PyJWT==2.10.1
httpx==0.28.1
apscheduler==3.10.4
resend==2.7.0
numpy==2.2.1
slowapi==0.1.9
